backend/requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# ML & Data Science
torch==2.2.0
torchvision==0.17.0
pillow==10.2.0
numpy==1.26.3
opencv-python==4.9.0.80
albumentations==1.3.1
timm==0.9.12
scikit-learn==1.4.0

# Explainability
grad-cam==1.4.8
matplotlib==3.8.2
seaborn==0.13.1

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1

# Google Cloud AI (for Gemini Pro)
google-cloud-aiplatform>=1.38.0

# Optional: LLM for narratives (uncomment if using)
# openai==1.10.0
# anthropic==0.8.1

# Development
pytest==7.4.4
httpx==0.26.0
black==24.1.1
ruff==0.1.14
